Changing the Gaia character in the erosion scene
