+++ src/3D-models/ErosionPlane.jsx
@@ -1,0 +1,36 @@
+import React, { useEffect } from "react";
+import { Suspense } from "react";
+import { useGLTF } from "@react-three/drei";
+
+function ErosionLoader() {
+  const gltf = useGLTF("3D-models/desert.glb");
+
+  useEffect(() => {
+    // Ensure all child meshes receive shadows
+    gltf.scene.traverse((child) => {
+      if (child.isMesh) {
+        child.receiveShadow = true; // Set to receive shadows
+      }
+    });
+  }, [gltf]);
+
+  return (
+    <primitive
+      receiveShadow
+      object={gltf.scene}
+      scale={20}
+      position={[-40, 0, 0]}
+    >
+      <shadowMaterial opacity={0.7} />
+    </primitive>
+  );
+}
+
+const ErosionPlane = () => {
+  return (
+    <Suspense fallback={null}>
+      <ErosionLoader />
+    </Suspense>
+  );
+};
+export default ErosionPlane;
