+++ src/3D-models/biodiversity/bio.css
@@ -1,0 +1,7 @@
+.parent {
+  color: blue; /* Set the text color for the parent */
+}
+
+.parent * {
+  background: inherit; /* All children will inherit the text color */
+}
