+++ src/3D-models/deforestation/Deforestation.css
@@ -1,0 +1,28 @@
+/* Desforestation.css */
+
+html,
+body,
+#root {
+  width: 100%;
+  height: 100%;
+  background: #77670a; /* Color de fondo */
+}
+
+canvas {
+  touch-action: none; /* Deshabilitar acciones táctiles */
+}
+
+.fixed-text {
+  position: absolute;
+  top: 100px;
+  right: 10px;
+  color: white;
+  font-size: 1.5rem;
+  z-index: 10;
+}
+
+/* Configura sombras del canvas */
+canvas {
+  display: block;
+  background: #77670a;
+}
