+++ src/pages/Home/Home.jsx
@@ -1,21 +1,21 @@
 import React from "react";
 import { Navbar } from "../../components/Navbar";
-import Gates from "../../Modelos3D/Maingates";
+import Gates from "../../3D-models/Maingates";
 import Welcome from "./Welcome"; // Asegúrate de que la ruta sea correcta
 import "./Home.css";
 import EnvironmentalCare from "./EnvironmentalCare";
 
 const LandingPage = () => {
-  return (
-    <>
-      <Navbar />
-      <div className="canvas-container">
-        <Gates />
-      </div>
-      <Welcome />
-      <EnvironmentalCare />
-    </>
-  );
+	return (
+		<>
+			<Navbar />
+			<div className="canvas-container">
+				<Gates />
+			</div>
+			<Welcome />
+			<EnvironmentalCare />
+		</>
+	);
 };
 
 export default LandingPage;
