+++ src/shared/3DModelLoader.tsx
@@ -1,0 +1,8 @@
+import { Html, useProgress } from "@react-three/drei";
+
+const Loader = () => {
+	const { progress } = useProgress();
+	return <Html center>{progress} % loaded</Html>;
+};
+
+export default Loader;
